# Track active series by ref instead of hash