--- a/pkg/ingester/activeseries/active_series.go
+++ b/pkg/ingester/activeseries/active_series.go
@@ -40,7 +40,7 @@ type seriesStripe struct {
 	oldestEntryTs atomic.Int64
 
 	mu             sync.RWMutex
-	refs           map[uint64][]seriesEntry
+	refs           map[uint64]seriesEntry
 	active         int   // Number of active entries in this stripe. Only decreased during purge or clear.
 	activeMatching []int // Number of active entries in this stripe matching each matcher of the configured Matchers.
 }
@@ -87,10 +87,10 @@ func (c *ActiveSeries) CurrentConfig() CustomTrackersConfig {
 }
 
 // UpdateSeries updates series timestamp to 'now'. Function is called to make a copy of labels if entry doesn't exist yet.
-func (c *ActiveSeries) UpdateSeries(series labels.Labels, fp uint64, now time.Time, labelsCopy func(labels.Labels) labels.Labels) {
-	stripeID := fp % numStripes
+func (c *ActiveSeries) UpdateSeries(series labels.Labels, ref uint64, now time.Time, labelsCopy func(labels.Labels) labels.Labels) {
+	stripeID := ref % numStripes
 
-	c.stripes[stripeID].updateSeriesTimestamp(now, series, fp, labelsCopy)
+	c.stripes[stripeID].updateSeriesTimestamp(now, series, ref, labelsCopy)
 }
 
 // purge removes expired entries from the cache.
@@ -137,13 +137,13 @@ func (s *seriesStripe) getTotalAndUpdateMatching(matching []int) int {
 	return s.active
 }
 
-func (s *seriesStripe) updateSeriesTimestamp(now time.Time, series labels.Labels, fingerprint uint64, labelsCopy func(labels.Labels) labels.Labels) {
+func (s *seriesStripe) updateSeriesTimestamp(now time.Time, series labels.Labels, ref uint64, labelsCopy func(labels.Labels) labels.Labels) {
 	nowNanos := now.UnixNano()
 
-	e := s.findEntryForSeries(fingerprint, series)
+	e := s.findEntryForSeries(ref, series)
 	entryTimeSet := false
 	if e == nil {
-		e, entryTimeSet = s.findOrCreateEntryForSeries(fingerprint, series, nowNanos, labelsCopy)
+		e, entryTimeSet = s.findOrCreateEntryForSeries(ref, series, nowNanos, labelsCopy)
 	}
 
 	if !entryTimeSet {
@@ -163,30 +163,28 @@ func (s *seriesStripe) updateSeriesTimestamp(now time.Time, series labels.Labels
 	}
 }
 
-func (s *seriesStripe) findEntryForSeries(fingerprint uint64, series labels.Labels) *atomic.Int64 {
+func (s *seriesStripe) findEntryForSeries(ref uint64, series labels.Labels) *atomic.Int64 {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 
 	// Check if already exists within the entries.
-	for _, entry := range s.refs[fingerprint] {
-		if labels.Equal(entry.lbs, series) {
-			return entry.nanos
-		}
+	entry, ok := s.refs[ref]
+	if !ok {
+		return nil
 	}
 
-	return nil
+	return entry.nanos
 }
 
-func (s *seriesStripe) findOrCreateEntryForSeries(fingerprint uint64, series labels.Labels, nowNanos int64, labelsCopy func(labels.Labels) labels.Labels) (*atomic.Int64, bool) {
+func (s *seriesStripe) findOrCreateEntryForSeries(ref uint64, series labels.Labels, nowNanos int64, labelsCopy func(labels.Labels) labels.Labels) (*atomic.Int64, bool) {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
 	// Check if already exists within the entries.
 	// This repeats findEntryForSeries(), but under write lock.
-	for _, entry := range s.refs[fingerprint] {
-		if labels.Equal(entry.lbs, series) {
-			return entry.nanos, false
-		}
+	entry, ok := s.refs[ref]
+	if ok {
+		return entry.nanos, false
 	}
 
 	matches := s.matchers.matches(series)
@@ -203,7 +201,7 @@ func (s *seriesStripe) findOrCreateEntryForSeries(fingerprint uint64, series lab
 		matches: matches,
 	}
 
-	s.refs[fingerprint] = append(s.refs[fingerprint], e)
+	s.refs[ref] = e
 
 	return e.nanos, true
 }
@@ -214,7 +212,7 @@ func (s *seriesStripe) clear() {
 	defer s.mu.Unlock()
 
 	s.oldestEntryTs.Store(0)
-	s.refs = map[uint64][]seriesEntry{}
+	s.refs = map[uint64]seriesEntry{}
 	s.active = 0
 	for i := range s.activeMatching {
 		s.activeMatching[i] = 0
@@ -227,7 +225,7 @@ func (s *seriesStripe) reinitialize(asm *Matchers) {
 	defer s.mu.Unlock()
 
 	s.oldestEntryTs.Store(0)
-	s.refs = map[uint64][]seriesEntry{}
+	s.refs = map[uint64]seriesEntry{}
 	s.active = 0
 	s.matchers = asm
 	s.activeMatching = resizeAndClear(len(asm.MatcherNames()), s.activeMatching)
@@ -247,55 +245,20 @@ func (s *seriesStripe) purge(keepUntil time.Time) {
 	s.activeMatching = resizeAndClear(len(s.activeMatching), s.activeMatching)
 
 	oldest := int64(math.MaxInt64)
-	for fp, entries := range s.refs {
-		// Since we do expect very few fingerprint collisions, we
-		// have an optimized implementation for the common case.
-		if len(entries) == 1 {
-			ts := entries[0].nanos.Load()
-			if ts < keepUntilNanos {
-				delete(s.refs, fp)
-				continue
-			}
-
-			s.active++
-			ml := entries[0].matches.len()
-			for i := 0; i < ml; i++ {
-				s.activeMatching[entries[0].matches.get(i)]++
-			}
-			if ts < oldest {
-				oldest = ts
-			}
+	for ref, entry := range s.refs {
+		ts := entry.nanos.Load()
+		if ts < keepUntilNanos {
+			delete(s.refs, ref)
 			continue
 		}
 
-		// We have more entries, which means there's a collision,
-		// so we have to iterate over the entries.
-		for i := 0; i < len(entries); {
-			ts := entries[i].nanos.Load()
-			if ts < keepUntilNanos {
-				entries = append(entries[:i], entries[i+1:]...)
-			} else {
-				if ts < oldest {
-					oldest = ts
-				}
-
-				i++
-			}
+		s.active++
+		ml := entry.matches.len()
+		for i := 0; i < ml; i++ {
+			s.activeMatching[entry.matches.get(i)]++
 		}
-
-		// Either update or delete the entries in the map
-		if cnt := len(entries); cnt == 0 {
-			delete(s.refs, fp)
-		} else {
-			s.active += cnt
-			for i := range entries {
-				ml := entries[i].matches.len()
-				for i := 0; i < ml; i++ {
-					s.activeMatching[entries[i].matches.get(i)]++
-				}
-			}
-
-			s.refs[fp] = entries
+		if ts < oldest {
+			oldest = ts
 		}
 	}