#### What this PR does

This PR changes the active series tracker to use series refs rather than hashes. This has the following advantages:
1. We no longer need to handle hash conflicts, since series refs are guaranteed unique by the tsdb.
2. It's now trivial to implement a `Contains(ref)` method on the active series tracker, which I will use to enhance the cardinality API so we can compute active series in the label values cardinality api. (This will be a follow up PR: #5135 )

I've run a local Mimir cluster and simulated series churn using the mimir-continuous-test tool, and active series appears to work just as before. 

For completeness, I ran the existing benchmarks and got the following:

<details><summary>Details</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/ingester/activeseries
                                                      │  before.txt   │              after.txt               │
                                                      │    sec/op     │    sec/op     vs base                │
ActiveSeriesTest_single_series/50-10                     156.6n ±  4%   140.1n ±  8%  -10.51% (p=0.000 n=10)
ActiveSeriesTest_single_series/100-10                    155.2n ±  7%   141.6n ±  5%   -8.73% (p=0.000 n=10)
ActiveSeriesTest_single_series/500-10                    158.8n ±  3%   142.7n ±  2%  -10.14% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=1_series=100000-10      33.64m ±  6%   20.89m ± 10%  -37.91% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=1_series=1000000-10     599.1m ±  7%   412.2m ± 10%  -31.19% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=100000-10    191.96m ± 11%   66.05m ± 13%  -65.59% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=1000000-10     2.915 ±  4%    1.482 ±  4%  -49.14% (p=0.000 n=10)
ActiveSeries_Active_once-10                              170.5µ ±  5%   142.9µ ±  4%  -16.16% (p=0.000 n=10)
ActiveSeries_Active_twice-10                             174.4µ ±  6%   149.5µ ±  1%  -14.27% (p=0.000 n=10)
geomean                                                  477.6µ         333.5µ        -30.18%

                                                      │   before.txt    │               after.txt                │
                                                      │      B/op       │     B/op      vs base                  │
ActiveSeriesTest_single_series/50-10                      0.000 ±  0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeriesTest_single_series/100-10                     0.000 ±  0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeriesTest_single_series/500-10                     0.000 ±  0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeries_UpdateSeries/rounds=1_series=100000-10     18.70Mi ±  0%     22.02Mi ± 0%  +17.78% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=1_series=1000000-10    227.4Mi ±  0%     336.7Mi ± 0%  +48.09% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=100000-10    18.70Mi ±  0%     22.02Mi ± 0%  +17.76% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=1000000-10   227.4Mi ±  0%     336.7Mi ± 0%  +48.11% (p=0.000 n=10)
ActiveSeries_Active_once-10                               78.00 ±  8%       69.00 ± 1%  -11.54% (p=0.000 n=10)
ActiveSeries_Active_twice-10                              80.50 ± 12%       71.50 ± 2%  -11.18% (p=0.000 n=10)
geomean                                                               ²                 +10.17%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                      │   before.txt   │               after.txt               │
                                                      │   allocs/op    │  allocs/op   vs base                  │
ActiveSeriesTest_single_series/50-10                     0.000 ±  0%      0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeriesTest_single_series/100-10                    0.000 ±  0%      0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeriesTest_single_series/500-10                    0.000 ±  0%      0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeries_UpdateSeries/rounds=1_series=100000-10     208.2k ±  0%     106.6k ± 0%  -48.79% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=1_series=1000000-10    2.021M ±  0%     1.021M ± 0%  -49.47% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=100000-10    208.2k ±  0%     106.6k ± 0%  -48.79% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=1000000-10   2.021M ±  0%     1.022M ± 0%  -49.46% (p=0.000 n=10)
ActiveSeries_Active_once-10                              3.000 ± 33%      3.000 ± 0%        ~ (p=0.087 n=10)
ActiveSeries_Active_twice-10                             5.000 ± 20%      4.000 ± 0%  -20.00% (p=0.011 n=10)
geomean                                                              ²                -27.76%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

```

</details> 

CPU usage is lower as expected, but allocated memory is slightly higher in `UpdateSeries` and lower in `Active`.

I've also consulted with @codesome about series refs to make sure this is a sane thing to do. Copying his response here (my questions in bold):

> **it seems like series refs in the head block are guaranteed to be unique across the lifetime of the process**

> That is right, but there is one small caveat. The same series can hold different series ref in the Head during the lifetime of the process, but at any given point, a series only has one ref. For example: {labels1} has ref=33, and then no samples came into that, so after the Head compaction the series was removed from Head. But then new samples started coming to it a little later, so a new series was created with new ref, but same labels, so now {labels1} has let's say ref=434 in the Head. 

> **but not necessarily after a restart**

> The series ID before and after restart can change. For example, taking the above example, if a restart happened and both series records for {labels1} existed in the WAL (ref 33 and ref 434), in this case the series created in the Head block for {labels1} will take the ref 33 here (since it comes first in the WAL). So after a restart, the ref for {labels1} changed from 434 to 33.

So it is possible that a single series may have multiple refs in two cases:
1. A series is removed from memory during head compaction. In this case, it will also be removed from the active series tracker first.
2. Ingesters restart and find multiple refs in the WAL. In this case, only the actual ref will be tracked in the active series tracker, since it starts empty and is filled by push requests over time.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/issues/5135

#### Checklist

- [X] Tests updated
- [N/A] Documentation added
- [N/A] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`