@colega 2023-06-02
Awesome, I wonder why we didn't think about this before? This totally makes sense.
I was curious so I ran the benchmarks for you:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester/activeseries
cpu: 11th Gen Intel(R) Core(TM) i7-11700K @ 3.60GHz
                                                      │   main.mem   │               ref.mem               │
                                                      │    sec/op    │   sec/op     vs base                │
ActiveSeriesTest_single_series/50-16                     29.59n ± 1%   38.08n ± 2%  +28.71% (p=0.000 n=10)
ActiveSeriesTest_single_series/100-16                    28.95n ± 2%   38.29n ± 1%  +32.26% (p=0.000 n=10)
ActiveSeriesTest_single_series/500-16                    28.82n ± 1%   37.38n ± 5%  +29.68% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=1_series=100000-16      23.97m ± 1%   20.37m ± 1%  -15.00% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=1_series=1000000-16     469.6m ± 2%   439.9m ± 1%   -6.33% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=100000-16    101.26m ± 4%   59.74m ± 2%  -41.01% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=1000000-16     1.937 ± 3%    1.346 ± 1%  -30.52% (p=0.000 n=10)
ActiveSeries_Active_once-16                              153.6µ ± 2%   136.9µ ± 1%  -10.87% (p=0.000 n=10)
ActiveSeries_Active_twice-16                             161.7µ ± 3%   144.3µ ± 0%  -10.73% (p=0.000 n=10)
geomean                                                  222.8µ        209.4µ        -6.00%

                                                      │    main.mem    │                ref.mem                 │
                                                      │      B/op      │     B/op      vs base                  │
ActiveSeriesTest_single_series/50-16                      0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeriesTest_single_series/100-16                     0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeriesTest_single_series/500-16                     0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeries_UpdateSeries/rounds=1_series=100000-16     18.70Mi ± 0%     22.95Mi ± 0%  +22.76% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=1_series=1000000-16    227.4Mi ± 0%     348.5Mi ± 0%  +53.25% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=100000-16    18.70Mi ± 0%     22.96Mi ± 0%  +22.78% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=1000000-16   227.4Mi ± 0%     348.5Mi ± 0%  +53.26% (p=0.000 n=10)
ActiveSeries_Active_once-16                               73.50 ± 5%       67.00 ± 9%   -8.84% (p=0.001 n=10)
ActiveSeries_Active_twice-16                              76.50 ± 3%       72.00 ± 6%   -5.88% (p=0.010 n=10)
geomean                                                              ²                 +13.14%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                      │   main.mem    │                ref.mem                │
                                                      │   allocs/op   │  allocs/op   vs base                  │
ActiveSeriesTest_single_series/50-16                     0.000 ± 0%      0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeriesTest_single_series/100-16                    0.000 ± 0%      0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeriesTest_single_series/500-16                    0.000 ± 0%      0.000 ± 0%        ~ (p=1.000 n=10) ¹
ActiveSeries_UpdateSeries/rounds=1_series=100000-16     208.2k ± 0%     108.2k ± 0%  -48.05% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=1_series=1000000-16    2.021M ± 0%     1.041M ± 0%  -48.51% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=100000-16    208.2k ± 0%     108.2k ± 0%  -48.05% (p=0.000 n=10)
ActiveSeries_UpdateSeries/rounds=10_series=1000000-16   2.021M ± 0%     1.041M ± 0%  -48.51% (p=0.000 n=10)
ActiveSeries_Active_once-16                              3.000 ± 0%      3.000 ± 0%        ~ (p=1.000 n=10)
ActiveSeries_Active_twice-16                             4.000 ± 0%      4.000 ± 0%        ~ (p=1.000 n=10)
geomean                                                             ²                -25.40%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

I wonder why the single series benchmarks are slower now, but the important is that the heavy ones are faster.