@Logiraptor 2023-06-02
@colega In the PR description I also had benchmark results, just hidden in a details block. Check it out, on my M1 mac the single series benchmarks are faster, and overall the change appears to be a bigger cpu difference than on your machine 🤔